pyproject.toml
[tool.poetry]
name = "tap-redshift"
version = "0.0.1"
description = "`tap-redshift` is a Singer tap for Redshift, built with the Meltano Singer SDK."
readme = "README.md"
authors = ["Shivam Kapoor <shivamk4r@gmail.com>"]
keywords = [
    "ELT",
    "Redshift",
]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "<3.12,>=3.7.1"
singer-sdk = { version="^0.32.0b2" }
fs-s3fs = { version = "^1.1.1", optional = true }
sqlalchemy-redshift = "^0.8.14"
redshift-connector = "^2.0.911"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
singer-sdk = { version="^0.32.0b2", extras = ["testing"] }

[tool.poetry.extras]
s3 = ["fs-s3fs"]

[tool.mypy]
python_version = "3.9"
warn_unused_configs = true
plugins = "sqlmypy"

[tool.ruff]
ignore = [
    "ANN101",  # missing-type-self
    "ANN102",  # missing-type-cls
]
select = ["ALL"]
src = ["tap_redshift"]
target-version = "py37"


[tool.ruff.flake8-annotations]
allow-star-arg-any = true

[tool.ruff.isort]
known-first-party = ["tap_redshift"]

[tool.ruff.pydocstyle]
convention = "google"

[build-system]
requires = ["poetry-core>=1.0.8"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
# CLI declaration
tap-redshift = 'tap_redshift.tap:TapRedshift.cli'
